로그인 유저 및 테이블 현황 서버 작업 중..
로그인 유저 및 테이블 현황 서버 작업 중..

+++ SP_Client/Assets/Script/Elements/ChatElt.cs
@@ -30,7 +30,7 @@ public class ChatElt : MonoBehaviour {
     const float ELT_MIN_WIDTH = 200f;
 	const float ELT_MIN_HEIGHT = 30f;
 
-    public void SetChatElt(byte person, byte customer, int tableNo, byte personCount, string time, string msg)
+    public void SetChatElt(byte person, byte customer, byte tableNo, byte personCount, string time, string msg)
 	{
         for (int i = 0; i < chatPersons.Length; i++)
             chatPersons[i].rtPerson.gameObject.SetActive(i == person);
